change SignInButton.tsx→GeneralButton.tsx & edit
