+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { signIn, useSession } from 'next-auth/react'
 import { createContext, useState } from 'react'
-import { SignInButton } from '../components/SignInButton'
+import { GeneralButton } from '../components/GeneralButton'
 import { Contents } from '../components/homeContents/Contents'
 
 export const HasKeyContext = createContext(
@@ -30,7 +30,9 @@ export default function Home() {
         <div className='flex items-center justify-center py-64'>
           <div>
             <h2>研究室のSlackアカウントによる認証を行なってください</h2>
-            <SignInButton clickFunction={slackSignIn} />
+            <div className='mt-14 text-center'>
+              <GeneralButton label='SIGN IN' clickFunction={slackSignIn} />
+            </div>
           </div>
         </div>
       )}
